backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pymongo==4.7.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
google-generativeai==0.5.4
PyMuPDF==1.24.3
pydantic[email]==2.7.1
httpx==0.27.0